+++ src/Http/Interactions/InteractsWithRequests.php
@@ -1,0 +1,20 @@
+<?php
+
+declare(strict_types=1);
+
+namespace SebastiaanLuca\Flow\Http\Interactions;
+
+use Illuminate\Contracts\Bus\Dispatcher;
+
+trait InteractsWithRequests
+{
+    /**
+     * @param \SebastiaanLuca\Flow\Http\Interactions\Interaction $interaction
+     *
+     * @return mixed
+     */
+    protected function interact(Interaction $interaction)
+    {
+        return app(Dispatcher::class)->dispatchNow($interaction);
+    }
+}