+++ CHANGELOG.md
@@ -8,7 +8,8 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ### Added
 
-- Added interactions
+- Added enforcer
+- Added interaction
 
 ### Removed